+++ src/main/java/com/linksteady/operate/service/TargetSplitAsyncService.java
@@ -1,11 +1,9 @@
 package com.linksteady.operate.service;
 
-import org.springframework.stereotype.Service;
-
 /**
  * 目标拆解异步任务
  */
 public interface TargetSplitAsyncService {
 
-     void targetSplit(int targetId);
+     void targetSplit(Long targetId);
 }
